pyproject.toml
[project]
name = "riven-memory"
version = "0.1.0"
description = "Memory API server for Riven - persistent vector memory with semantic search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.5.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "aiofiles>=23.0.0",
    "requests>=2.28.0",
    "pydantic>=2.0.0",
]

[project.scripts]
riven-memory = "server:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
